정지웅/25.12.03/휴일/근무형태 기준 관리 거의 완료

+++ src/main/java/store/yd2team/insa/service/HldyVO.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public class HldyVO {
 
     /** 휴일번호 (PK, NUMBER(10)) */
-    private Long hldyNo;
+    private Integer hldyNo;
 
     /** 휴일명 */
     private String hldyNm;
